# src/backend/services/calendar-service/docs/swagger.json
{"swagger": "2.0", "basePath": "/rest/api/v1", "paths": {"/authorization/": {"get": {"responses": {"400": {"description": "Error", "schema": {"$ref": "#/definitions/Calendar%20service%20authorization%20response%20-%20error"}}, "200": {"description": "OK", "schema": {"$ref": "#/definitions/Calendar%20service%20authorization%20response%20-%20success"}}}, "description": "Authorize with google.", "operationId": "get_authorize", "tags": ["authorization"]}}, "/authorization/clear": {"get": {"responses": {"400": {"description": "Error", "schema": {"$ref": "#/definitions/Clear%20response%20-%20error"}}, "200": {"description": "OK", "schema": {"$ref": "#/definitions/Clear%20response%20-%20success"}}}, "description": "Clear all authentication data, if you want to completely remove the permissions, call /authorization/revoke first.", "operationId": "get_clear", "tags": ["authorization"]}}, "/authorization/oauth2callback": {"get": {"responses": {"400": {"description": "Error", "schema": {"$ref": "#/definitions/Callback%20response%20-%20error"}}, "200": {"description": "OK", "schema": {"$ref": "#/definitions/Callback%20response%20-%20success"}}}, "description": "Callback to be called after authentication.", "operationId": "get_callback", "tags": ["authorization"]}}, "/authorization/revoke": {"get": {"responses": {"400": {"description": "Error", "schema": {"$ref": "#/definitions/Revoke%20response%20-%20error"}}, "200": {"description": "OK", "schema": {"$ref": "#/definitions/Revoke%20response%20-%20success"}}}, "description": "Revoke the permissions.", "operationId": "get_revoke", "tags": ["authorization"]}}, "/events/": {"get": {"responses": {"400": {"description": "Error"}, "200": {"description": "OK"}}, "description": "Get events with specified date, if no date is specified, the current date is used.", "operationId": "get_events", "tags": ["events"]}}, "/events/{date}": {"parameters": [{"name": "date", "in": "path", "required": true, "type": "string"}], "get": {"responses": {"400": {"description": "Error"}, "200": {"description": "OK"}}, "description": "Get events with specified date, if no date is specified, the current date is used.", "operationId": "get_events", "tags": ["events"]}}, "/swagger/swagger.json": {"get": {"responses": {"200": {"description": "OK"}}, "description": "Generate a swagger.json file.", "operationId": "get_swagger", "tags": ["swagger"]}}, "/swagger/swagger.yml": {"get": {"responses": {"200": {"description": "OK"}}, "description": "Generate a swagger.yml file.", "operationId": "get_swagger", "tags": ["swagger"]}}}, "info": {"title": "REST-API for Calendar-Service", "version": "1.0.0", "description": "Service to receive appointments from google calendar"}, "produces": ["application/json"], "consumes": ["application/json"], "tags": [{"name": "swagger", "description": "Swagger APIs"}, {"name": "events", "description": "Google calendar events APIs"}, {"name": "authorization", "description": "Google authentication APIs"}], "definitions": {"Calendar service authorization response - error": {"properties": {"error": {"type": "string"}}, "type": "object"}, "Calendar service authorization response - success": {"properties": {"authorization_url": {"type": "string"}}, "type": "object"}, "Callback response - error": {"properties": {"error": {"type": "string"}}, "type": "object"}, "Callback response - success": {"properties": {"message": {"type": "string"}}, "type": "object"}, "Revoke response - error": {"properties": {"error": {"type": "string"}}, "type": "object"}, "Revoke response - success": {"properties": {"message": {"type": "string"}}, "type": "object"}, "Clear response - error": {"properties": {"error": {"type": "string"}}, "type": "object"}, "Clear response - success": {"properties": {"message": {"type": "string"}}, "type": "object"}}, "responses": {"ParseError": {"description": "When a mask can't be parsed"}, "MaskError": {"description": "When any error occurs on mask"}}}